Update nextjs.mdx
Ran into this issue with the docs when implementing this in our Next.js app. The docs say to import `initAuth` from `@propelauth/express` and then to reference `validateAccessTokenAndGetUser` from the resulting object, but `validateAccessTokenAndGetUser` does not exist on that object. I eventually figured out that this was a typo, and that I was meant to use `initBaseAuth` from `@propelauth/node`