Refactor the docs to be use-case oriented
The top level of the docs right now is about frameworks.
You choose a framework and then get a lot of instructions
about using PropelAuth with that framework.

There are some benefits to that approach, but it also feels
like it we can do a better job explaining things if we have
shorter guides that are broken down into the outcomes they provide.

This is draft that does that, specifically:
- Break up the overview into sections, including a key concept sections
- Introduce the hosted pages, FE, and BE as key concepts
- Adds a quickstart guide
- Turns the existing integration guides into more focused guides

This also only does one language so far. Want to get feedback before doing
all of them.